+++ package.lisp
@@ -1,3 +1,5 @@
 (defpackage #:com.aragaer.pa-brain
   (:use #:common-lisp)
-  (:export :react))
+  (:export :react)
+  (:export :handle-message)
+  (:export :*active-user*))